requirements.txt
streamlit
python-dotenv
langchain==0.1.16
langchain-community==0.0.36
langchain-openrouter==0.0.5
faiss-cpu
sentence-transformers
